requirements.txt
litellm 
opentelemetry-sdk 
opentelemetry-exporter-gcp-trace
